Added more testing for tweets

+++ routes/tweetRouter.js
@@ -7,6 +7,17 @@ const tweetRouter = express.Router();
 // Create tweet
 // If req.body contains parentId then it's a retweet
 tweetRouter.post("/", expressAsyncHandler( async(req, res) => {
+    if (req.body.parentId != null) {
+        const currentTweet = await Tweet.findById(req.body.parentId)
+            .catch(err => {
+                res.status(500).send({success: false, error: err});
+                return;
+            });
+        if (currentTweet == null) {
+            res.status(406).send({success: false, error: "Parent tweet doesn't exist"});
+            return;
+        }
+    }
     const newTweet = new Tweet(req.body);
     const saveTweet = await newTweet.save()
         .catch(err => {
@@ -18,7 +29,7 @@ tweetRouter.post("/", expressAsyncHandler( async(req, res) => {
 
 // Create subtweet (Cannot contain parentId)
 tweetRouter.post("/subtweet/:id", expressAsyncHandler( async(req, res) => {
-    if (req.body.parentId && req.body.parentId.length > 0) {
+    if (req.body.parentId != null && req.body.parentId.length > 0) {
         res.status(403).send({success: false, error: "Cannot be both subtweet and retweet"});
     } else {
         // Find tweet we're subtweeting
@@ -27,23 +38,27 @@ tweetRouter.post("/subtweet/:id", expressAsyncHandler( async(req, res) => {
                 res.status(500).send({success: false, error: err});
                 return;
             });
-        // Create new tweet body
-        const newTweet = new Tweet(req.body);
-        const saveTweet = await newTweet.save()
-            .catch(err => {
-                res.status(500).send({success: false, error: err});
-                return;
-            });
-        const updatedTweet = await currentTweet.updateOne({
-            $push : {subTweets: saveTweet._id}
-        })
-            .catch(err => {
-                res.status(500).send({success: false, error: err});
-                return;
+        if (currentTweet !== null) {
+            // Create new tweet body
+            const newTweet = new Tweet(req.body);
+            const saveTweet = await newTweet.save()
+                .catch(err => {
+                    res.status(500).send({success: false, error: err});
+                    return;
+                });
+            const updatedTweet = await currentTweet.updateOne({
+                $push : {subTweets: saveTweet._id}
             })
-        res.status(200).send({success: true, message: "Subtweet successful"});
-    }
+                .catch(err => {
+                    res.status(500).send({success: false, error: err});
+                    return;
+                })
+            res.status(200).send({success: true, message: "Subtweet successful"});
+        } else {
+            res.status(406).send({success: false, error: "Tweet doesn't exist"})
+        }
 
+    }
 }));
 
 // Get posts
@@ -56,7 +71,12 @@ tweetRouter.get("/:id", expressAsyncHandler( async( req, res) => {
             res.status(500).send({success: false, error: err});
             return;
         });
-    res.status(200).send({success: true, message: tweet});
+    if (tweet !== null) {
+        res.status(200).send({success: true, message: tweet});
+    } else {
+        res.status(406).send({success: false, error: "Tweet doesn't exist"})
+    }
+    
 }));
 
 // GET subTweets for loading comments (By tweetId)
@@ -64,12 +84,16 @@ tweetRouter.get("/:id", expressAsyncHandler( async( req, res) => {
 tweetRouter.get("/comments/:id", expressAsyncHandler( async(req, res) => {
     try {
         const currentTweet = await Tweet.findById(req.params.id);
-        const tweets = await Promise.all(
-            currentTweet.subTweets.map(async (tweetId) => {
-                return Tweet.findById(tweetId);
-            })
-        );
-        res.status(200).send({success: true, message: tweets});
+        if (currentTweet !== null) {
+            const tweets = await Promise.all(
+                currentTweet.subTweets.map(async (tweetId) => {
+                    return Tweet.findById(tweetId);
+                })
+            );
+            res.status(200).send({success: true, message: tweets});
+        } else {
+            res.status(406).send({success: false, error: "Tweet doesn't exist"})
+        }
     } catch (err) {
         res.status(500).send({success: false, error: err});
     }
@@ -80,13 +104,17 @@ tweetRouter.get("/comments/:id", expressAsyncHandler( async(req, res) => {
 tweetRouter.get("/find/timeline", expressAsyncHandler( async(req, res) => {
     try {
         const currentUser = await User.findById(req.body.id);
-        const tweets = await Tweet.findById({ userId: currentUser._id });
-        const followingTweets = await Promise.all(  // Promise all takes list of promises and returns list of results
-            currentUser.following.map(async (followId) => {  // use .map to loop through all userId in following array (followId), then create promise (Post.find) to fetch that users tweets
-                return Tweet.find({ userId: followId });
-            })
-        );
-        res.status(200).send({success: true, message: tweets.concat(...followingTweets)});
+        const tweet = await Tweet.findById({ userId: currentUser._id });
+        if (tweet !== null) {
+            const followingTweets = await Promise.all(  // Promise all takes list of promises and returns list of results
+                currentUser.following.map(async (followId) => {  // use .map to loop through all userId in following array (followId), then create promise (Post.find) to fetch that users tweets
+                    return Tweet.find({ userId: followId });
+                })
+            );
+            res.status(200).send({success: true, message: tweet.concat(...followingTweets)});
+        } else {
+            res.status(406).send({success: false, error: "Tweet doesn't exist"});
+        }
     } catch (err) {
         res.status(500).send({success: false, error: err});
     }
@@ -100,12 +128,17 @@ tweetRouter.put("/:id", expressAsyncHandler( async(req, res) => {
             res.status(500).send({success: false, error: err});
             return;
         });
-    if (tweet.userId === req.body.id) {
-        await tweet.updateOne({ $set: req.body });
-        res.status(200).send({success: true, message: "The tweet has been updated" });
+    if (tweet !== null) {
+        if (tweet.userId === req.body.id) {
+            await tweet.updateOne({ $set: req.body });
+            res.status(200).send({success: true, message: "The tweet has been updated" });
+        } else {
+            res.status(401).send({success: false, error: "You can only update your own tweets" });
+        }
     } else {
-        res.status(401).send({success: false, error: "You can only update your own tweets" });
+        res.status(406).send({success: false, error: "Tweet doesn't exist"})
     }
+
 }));
 
 // Add UPDATE to add a subtweet (Comment should create own tweet then add that tweets ID to subTweets array)
@@ -124,12 +157,17 @@ tweetRouter.delete("/:id", expressAsyncHandler( async(req, res) => {
             res.status(500).send({success: false, error: err});
             return;
         });
-    if (tweet.userId === req.body.id) {  // Again if admin feat added check || req.body.isAdmin
-        await tweet.deleteOne();
-        res.status(200).send({success: true, message: "The tweet has been deleted"});
+    if (tweet !== null) {
+        if (tweet.userId === req.body.id) {  // Again if admin feat added check || req.body.isAdmin
+            await tweet.deleteOne();
+            res.status(200).send({success: true, message: "The tweet has been deleted"});
+        } else {
+            res.status(401).send({success: false, error: "You can only delete your own tweets" });
+        }
     } else {
-        res.status(401).send({success: false, error: "You can only delete your own tweets" });
+        res.status(406).send({success: false, error: "Tweet doesn't exist"})
     }
+
 }));
 
 // Like & Dislike tweet
@@ -140,13 +178,18 @@ tweetRouter.put("/like/:id", expressAsyncHandler( async(req, res) => {
             res.status(500).send({success: false, error: err});
             return;
         });
-    if (!tweet.likes.includes(req.body.id)) {  // Tweet not liked, add id to likes array
-        await tweet.updateOne({ $push: { likes: req.body.id }});
-        res.status(200).send({success: true, message: "Tweet has been liked"});
-    } else {  // Tweet liked, remove id from likes array
-        await tweet.updateOne({ $pull: { likes: req.body.id }});
-        res.status(200).send({success: true, message: "Tweet has been unliked"});
+    if (tweet !== null) {
+        if (!tweet.likes.includes(req.body.id)) {  // Tweet not liked, add id to likes array
+            await tweet.updateOne({ $push: { likes: req.body.id }});
+            res.status(200).send({success: true, message: "Tweet has been liked"});
+        } else {  // Tweet liked, remove id from likes array
+            await tweet.updateOne({ $pull: { likes: req.body.id }});
+            res.status(200).send({success: true, message: "Tweet has been unliked"});
+        }
+    } else {
+        res.status(406).send({success: false, error: "Tweet doesn't exist"});
     }
+
 }));
